requirements/base.txt
asgiref==3.2.7
backcall==0.1.0
decorator==4.4.2
Django==3.0.7
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.9
cryptography
exponent_server_sdk
python-dateutil
djangorestframework-simplejwt
